:art: Extract package interface to separate file
