+++ lib/main.js
@@ -1,0 +1,17 @@
+import configSchema from './config-schema';
+import {toggleQuotes} from './toggle-quotes';
+
+export default {
+  config: configSchema,
+
+  activate: () => {
+    atom.commands.add('atom-text-editor', 'toggle-quotes:toggle', () => {
+      let editor = atom.workspace.getActiveTextEditor();
+      if (editor) {
+        toggleQuotes(editor);
+      }
+    });
+  },
+
+  toggleQuotes: toggleQuotes
+};
