+++ lib/toggle-quotes.js
@@ -2,7 +2,7 @@
 
 import configSchema from './config-schema';
 
-const toggleQuotes = (editor) => {
+export const toggleQuotes = (editor) => {
   editor.transact(() => {
     for (let cursor of editor.getCursors()) {
       let position = cursor.getBufferPosition();
@@ -73,19 +73,4 @@ const toggleQuoteAtPosition = (editor, position) => {
   newText = prefix + nextQuoteCharacter + newText.slice(1 + prefix.length, -1) + nextQuoteCharacter;
 
   editor.setTextInBufferRange(range, newText);
-};
-
-export default {
-  config: configSchema,
-
-  activate: () => {
-    atom.commands.add('atom-text-editor', 'toggle-quotes:toggle', () => {
-      let editor = atom.workspace.getActiveTextEditor();
-      if (editor) {
-        toggleQuotes(editor);
-      }
-    });
-  },
-
-  toggleQuotes: toggleQuotes
 };
