+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "toggle-quotes",
   "version": "0.11.0",
-  "main": "./lib/toggle-quotes",
+  "main": "./lib/main",
   "description": "Quickly toggle between single and double quotes",
   "repository": "https://github.com/atom/toggle-quotes",
   "license": "MIT",
